test(ui): split the appearance tests into their own file
Theme, accent, background ownership, and colour mode each have persistence and
precedence tests, which together were the second largest group in the file.

Tests move unchanged.
